Move source codes under `src` directory
This makes it easier to add e.g. new directories without having
to update tooling every time.
Next.js supports this change by default:
https://nextjs.org/docs/advanced-features/src-directory
