+++ Dockerfile
@@ -2,9 +2,8 @@ FROM node:15.8.0-alpine3.12 AS build
 WORKDIR /app
 COPY package.json yarn.lock ./
 RUN yarn install --frozen-lockfile
-COPY ./pages ./pages
+COPY ./src ./src
 COPY ./public ./public
-COPY ./styles ./styles
 COPY tsconfig.json  ./
 ARG NEXT_PUBLIC_GIT_HASH=unknown
 RUN yarn build
